+++ Sources/TABResourceLoader/Protocols/ResourceServiceType.swift
@@ -30,24 +30,27 @@ import Foundation
 
 /// Defines a type that can be cancelled
 public protocol Cancellable: class {
+  /// Cancels this task
   func cancel()
 }
 
 /// Defines a type that can fetch resources
 public protocol ResourceServiceType {
+  /// Defines the type of resource that this service fetches
   associatedtype Resource: ResourceType
 
-  /**
-   Designated initialzer for constructing a ResourceServiceType
-   */
-  init(session: URLSessionType)
+  /// Defines the type of result that this service returns
+  associatedtype ResultType
 
-  /**
-   Fetch a resource
+  /// Designated initialzer for constructing a ResourceServiceType
+  init()
 
-   - parameter resource:   The resource to fetch
-   - parameter completion: A completion handler called with a Result type of the fetching computation
-   */
+  /// Fetch a resource
+  ///
+  /// - Parameters:
+  ///   - resource: The resource to fetch
+  ///   - completion: A completion handler called with a Result type of the fetching computation
+  /// - Returns: A cancellable type if it's possible to cancel the task created
   @discardableResult
-  func fetch(resource: Resource, completion: @escaping (Result<Resource.Model>) -> Void) -> Cancellable?
+  func fetch(resource: Resource, completion: @escaping (ResultType) -> Void) -> Cancellable?
 }