+++ Tests/TABResourceLoaderTests/Protocols/NetworkDataResourceServiceURLSessionTests.swift
@@ -22,13 +22,13 @@ class NetworkDataResourceServiceURLSessionTests: XCTestCase { //swiftlint:disabl
   struct MockResource: NetworkResourceType, DataResourceType {
     typealias Model = String
     var url = URL(string: "test")!
-    func result(from data: Data) -> Result<Model> {
-      return .success("done")
+    func model(from data: Data) throws -> Model {
+      return "done"
     }
   }
 
   func test_initWithURLSession() {
-    let service = NetworkDataResourceService<MockResource>(session: MockSessionToTestExposed())
+    let service = GenericNetworkDataResourceService<MockResource>(session: MockSessionToTestExposed())
     let resource = MockResource()
 
     var hasFetched = false